Assignments/Task-3/22671A7326/requirements.txt
streamlit
opencv-python
numpy
scikit-image
pillow